Remove user/group creation (it's useless)
Currently the UID/GID is optionally set by the host, and it can be set to any arbitrary value (not just 1000). Since 1. the container itself makes no use of the created user/group and 2. the dnsdist packages does not require a home directory, there is really no point of creating the user/group. It would make sense if we were to forcefully downgrade the user by using docker's USER command - but we don't. By default package runs as root and the user can optionally change the UID/GID by setting proper docker arguments.